Split up GroupKeyword
